fix: 번역 파일 통합 및 페이지 이동 시 언어 유지 개선
