+++ features/my-pokedex/components/PokedexGrid.tsx
@@ -3,7 +3,7 @@
 import { useRef, useEffect } from "react";
 import Link from "next/link";
 import { PokemonInDex } from "@/features/my-pokedex/usecases/getMyPokedex";
-import { useTranslation } from "@/features/shared/utils/translate/useTranslation";
+import { useTranslation } from "@/features/shared/utils/translate/useLanguage";
 
 interface PokedexGridProps {
   pokemons: PokemonInDex[];
