+++ features/shared/utils/translate/useLanguage.tsx
@@ -1,6 +1,10 @@
 "use client";
 
-import { createContext, useContext, useState, ReactNode } from "react";
+import { createContext, useContext, useState, useEffect, ReactNode } from "react";
+import { usePathname } from "next/navigation";
+import { ko } from "./languages/ko";
+import { en } from "./languages/en";
+import { ja } from "./languages/ja";
 
 export type Language = "ko" | "en" | "ja";
 
@@ -10,6 +14,39 @@ interface LanguageContextType {
 }
 
 const LanguageContext = createContext<LanguageContextType | undefined>(undefined);
+
+// ===== Translation Types & Functions =====
+
+export const translations = {
+  ko,
+  en,
+  ja,
+} as const;
+
+type NestedKeyOf<T> = T extends object
+  ? {
+      [K in keyof T]: K extends string
+        ? T[K] extends object
+          ? `${K}.${NestedKeyOf<T[K]>}`
+          : K
+        : never;
+    }[keyof T]
+  : never;
+
+export type TranslationKey = NestedKeyOf<typeof translations.ko>;
+
+export function getTranslation(language: Language, key: TranslationKey): string {
+  const keys = key.split(".");
+  let value: any = translations[language];
+
+  for (const k of keys) {
+    value = value?.[k];
+  }
+
+  return value || key;
+}
+
+// ===== Language Context & Provider =====
 
 // 브라우저 언어 감지 함수
 function detectBrowserLanguage(): Language {
@@ -35,13 +72,47 @@ function getInitialLanguage(): Language {
 }
 
 export function LanguageProvider({ children }: { children: ReactNode }) {
+  const pathname = usePathname();
   const [language, setLanguageState] = useState<Language>(() => getInitialLanguage());
+  const [mounted, setMounted] = useState(false);
+
+  // 클라이언트 마운트 확인 및 localStorage 동기화
+  useEffect(() => {
+    setMounted(true);
+    const savedLanguage = localStorage.getItem("language") as Language;
+    if (savedLanguage && ["ko", "en", "ja"].includes(savedLanguage)) {
+      if (savedLanguage !== language) {
+        setLanguageState(savedLanguage);
+      }
+    }
+  }, []);
+
+  // pathname 변경 시 localStorage 재확인 (뒤로가기 대응)
+  useEffect(() => {
+    if (!mounted) return;
+
+    const savedLanguage = localStorage.getItem("language") as Language;
+    if (savedLanguage && ["ko", "en", "ja"].includes(savedLanguage)) {
+      if (savedLanguage !== language) {
+        setLanguageState(savedLanguage);
+      }
+    }
+  }, [pathname, mounted]);
 
   // 언어 변경 함수
   const setLanguage = (lang: Language) => {
     setLanguageState(lang);
     localStorage.setItem("language", lang);
   };
+
+  // SSR 중에는 기본값 사용
+  if (!mounted) {
+    return (
+      <LanguageContext.Provider value={{ language, setLanguage }}>
+        {children}
+      </LanguageContext.Provider>
+    );
+  }
 
   return (
     <LanguageContext.Provider value={{ language, setLanguage }}>
@@ -56,4 +127,16 @@ export function useLanguage() {
     throw new Error("useLanguage must be used within a LanguageProvider");
   }
   return context;
+}
+
+// ===== Translation Hook =====
+
+export function useTranslation() {
+  const { language } = useLanguage();
+
+  const translate = (key: TranslationKey): string => {
+    return getTranslation(language, key);
+  };
+
+  return { translate, language };
 }
